chore(): deprecate pre-V5 authentication workflows
