+++ src/rest-client-v5.ts
@@ -272,7 +272,7 @@ export class RestClientV5 extends BaseRestClient {
    */
 
   getClientType() {
-    return REST_CLIENT_TYPE_ENUM.v3;
+    return REST_CLIENT_TYPE_ENUM.v5;
   }
 
   async fetchServerTime(): Promise<number> {
